==> research/03_prepare_base_model.ipynb <==
import os 


pwd%


os.chdir("../")

pwd%


lazem nzid l2_reg w nchouf kifech bich naddi il shape 

from dataclasses import dataclass
from pathlib import Path
@dataclass(frozen=True)
class PrepareBaseModelConfig:
    root_dir:Path
    full_model_path:Path
    shape_path:Path
    lstm_units_1:int 
    lstm_units_2 :int 
    dropout_rate_1:float
    dropout_rate_2:float
    l2_reg_1:float
    l2_reg_2:float


from src.LSTM_BTC_Prediction.constants  import *
from src.LSTM_BTC_Prediction.utils.common import read_yaml,create_directories

class ConfigurationManager:
    def __init__(
            self,
            config_filepath=CONFIG_FILE_PATH,
            params_filepath=PARAMS_FILE_PATH):
        self.config=read_yaml(config_filepath) 
        self.params=read_yaml(params_filepath)

        create_directories([self.config.artifacts_root])

    def get_prepare_base_model_config(self) -> PrepareBaseModelConfig:
        config=self.config.prepare_base_model

        create_directories([config.root_dir])  

        prepare_base_model_config=PrepareBaseModelConfig(
            root_dir= config.root_dir,
            full_model_path = config.full_model_path,
            shape_path = config.shape_path,
            lstm_units_1= self.params.LSTM_UNITS_1,
            lstm_units_2 = self.params.LSTM_UNITS_2,
            dropout_rate_1= self.params.DROPOUT_RATE_1,
            dropout_rate_2 = self.params.DROPOUT_RATE_2 ,
            l2_reg_1=self.params.L2_REG_1,
            l2_reg_2=self.params.L2_REG_2  )
        
        return prepare_base_model_config

import tensorflow as tf
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Bidirectional, LSTM, Dropout, Dense
from tensorflow.keras.regularizers import L2
from src.LSTM_BTC_Prediction import logger 
import numpy as np 


class PrepareBaseModel:

    def __init__(self, config: PrepareBaseModelConfig):
        self.config = config

    def get_shape(self):
        try:
            self.trainX = np.load(self.config.shape_path) 
            logger.info("Downloading trainX successful")
        except Exception as e:
            logger.error(f"Error occurred during downloading trainX: {e}")

    def get_base_model(self):
        try:
            self.get_shape()  # Call get_shape to ensure trainX is loaded before building the model

            self.model = Sequential()
            self.model.add(Bidirectional(LSTM(units=self.config.lstm_units_1, return_sequences=True, kernel_regularizer=L2(self.config.l2_reg_1)),
                                         input_shape=(self.trainX.shape[1], self.trainX.shape[2])))
            self.model.add(Dropout(self.config.dropout_rate_1))
            self.model.add(Bidirectional(LSTM(units=self.config.lstm_units_2, kernel_regularizer=L2(self.config.l2_reg_2))))
            self.model.add(Dropout(self.config.dropout_rate_2))
            self.model.add(Dense(units=1))
            self.model.summary()
            logger.info("Compiling model completed")
            self.save_model(model=self.model, path=self.config.full_model_path)
            logger.info('model saved')
        except Exception as e:
            logger.error(f"Error occurred during building the base model: {e}")

    @staticmethod
    def save_model(path: Path, model: tf.keras.Model):
        model.save(path)


try:
    config = ConfigurationManager()
    prepare_base_model_config = config.get_prepare_base_model_config()
    prepare_base_model = PrepareBaseModel(config=prepare_base_model_config)
    prepare_base_model.get_shape()
    prepare_base_model.get_base_model() 
except Exception as e:
    raise e
